+++ src/components/alpaca/chart-position-modal.tsx
@@ -5,9 +5,10 @@ import { Dialog, DialogContent, DialogTitle, IconButton } from '@mui/material';
 import { basicFetch } from '@/app/lib/fetchFunctions';
 import { Position } from '@/models/strategy/position';
 import CircularLoader from '../common/loader';
-import PositionChart from './history-charts/position-chart';
+import PositionChartBreakout from './history-charts/position-chart-breakout';
 import { format } from 'date-fns';
 import CloseIcon from '@mui/icons-material/Close';
+import { StrategyEnum } from '@/models/strategy/enums';
 
 interface ChartPositionModalProps {
     isOpen: boolean;
@@ -31,13 +32,22 @@ const ChartPositionModal: React.FC<ChartPositionModalProps> = (params) => {
     const getChartData = async () => {
         setLoading(true);
 
-        const prevLowStamp = new Date(params.position.prevLowStamp);
-        const prevHighStamp = new Date(params.position.prevHighStamp);
-        const dateOpened = (prevLowStamp > prevHighStamp) ? new Date(prevHighStamp) : new Date(prevLowStamp);
-        dateOpened.setDate(dateOpened.getDate() - 1);
+        let dateOpened = new Date(params.position.stampOpened);
+
+        if (params.position.strategyType === StrategyEnum.Breakout) {
+
+            var breakoutParams = JSON.parse(params.position.strategyParams);
+            const prevLowStamp = new Date(breakoutParams.PrevLowStamp);
+            const prevHighStamp = new Date(breakoutParams.PrevHighStamp);
+            dateOpened = (prevLowStamp > prevHighStamp) ? new Date(prevHighStamp) : new Date(prevLowStamp);
+
+        }
+
+        dateOpened.setDate(dateOpened.getDate() - 10);
+
         const formattedDateOpened = format(new Date(dateOpened), 'yyyy-MM-dd');
         const dateClosed = new Date(params.position.stampClosed);
-        dateClosed.setDate(dateClosed.getDate() + 1);
+        dateClosed.setDate(dateClosed.getDate() + 10);
         const formattedDateClosed = format(new Date(dateClosed), 'yyyy-MM-dd');
 
         console.log('dateOpened:', formattedDateOpened, 'dateClosed:', formattedDateClosed);
@@ -63,7 +73,10 @@ const ChartPositionModal: React.FC<ChartPositionModalProps> = (params) => {
                     </div>
                 </DialogTitle>
                 <DialogContent>
-                    <div className='w-[900px] h-[600px]'> {loading ? <CircularLoader /> : <PositionChart data={chartData} position={params.position} />}</div>
+                    {params.position.strategyType === StrategyEnum.Breakout && <>
+                        <div className='w-[900px] h-[600px]'> {loading ? <CircularLoader /> : <PositionChartBreakout data={chartData} position={params.position} />}</div>
+                    </>}
+
                 </DialogContent>
             </Dialog>
         </>