+++ src/models/strategy/breakout-settings-model.ts
@@ -1,0 +1,7 @@
+import { StopLossTypeEnum, BreakoutPeriodEnum } from "./enums";
+
+
+export interface BreakoutSettingsModel {
+  stopLossType: StopLossTypeEnum;
+  breakoutPeriod: BreakoutPeriodEnum;
+}