+++ src/models/strategy/sma-model.ts
@@ -1,0 +1,6 @@
+import { StrategySettingsModel } from "./strategy-settings-model";
+
+export interface SMAModel extends StrategySettingsModel {
+  shortPeriod: number;
+  longPeriod: number;
+}